requirements.txt
setuptools
requests
discord.py
python-dotenv
asyncpg